# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Berim API",
    "description": "<div style='direction: rtl'>به کمک نرم‌افزار بریم با دوستان خود برنامه بیرون رفتن هماهنگ کنید. <br/> این سند از چهار بخش تشکیل شده که در هر بخش توابع مربوط به همان بخش به همراه ورودی ها و خروجی های آن امده است. این بخش ها عبارتند از کاربر، مکان‌ها، دوستان و گفتگوها. سعی شده توضیحات هر تابع درون خود تابع داده شود و سند گویا باشد. به کمک شما دوستان کم کم این سند رو تکمیل خواهیم کرد. <br/> از دیگر مزیت های این سیستم قابلیت تست ای پی ای هست که در توابع زیر خواهید دید که میتونیم ورودی های هر تابع رو وارد کنیم و نتیجه خروجی رو ببینیم که مطمئنا خیلی کمکمون خواهد کرد. <br/> با سپاس فراوان :)</div>",
    "version": "1.0.0"
  },
  "produces": ["application/json"],
  "host": "localhost:8080",
  "basePath": "/v1",
  "paths": {
  "/user/resend_activation_code": {
    
    "post": {
      "summary": "send an activation code to user again",
      "description":"each user need to activate his/her account so we send an activation code to user that user should enter it in client after signup",
      "x-swagger-router-controller": "User",
      "tags": ["User"],
      "operationId": "userResendActivationCodePost",
      "parameters": [
        {
  "name" : "phone_number",
  "in" : "query",
  "description" : "user phone number.",
  "required" : true,
  "type" : "number",
  "format" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/DoneResponse"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/user/get_profile": {
    
    "get": {
      "summary": "get user Profile",
      "description":"The User Profile endpoint returns information about the user that has authorized with the application.",
      "x-swagger-router-controller": "User",
      "tags": ["User"],
      "operationId": "userGetProfileGet",
      "parameters": [
        {
  "name" : "phone_number",
  "in" : "query",
  "description" : "user phone number.",
  "required" : true,
  "type" : "number",
  "format" : "string"
},
        {
  "name" : "password",
  "in" : "query",
  "description" : "user password.",
  "required" : true,
  "type" : "string",
  "format" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "Profile information for a user",
  "schema" : {
    "$ref" : "#/definitions/User"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/user/signin": {
    
    "post": {
      "summary": "login user",
      "description":"get phone number and password and check user login",
      "x-swagger-router-controller": "User",
      "tags": ["User"],
      "operationId": "userSigninPost",
      "parameters": [
        {
  "name" : "phone_number",
  "in" : "query",
  "description" : "user phone number.",
  "required" : true,
  "type" : "number",
  "format" : "string"
},
        {
  "name" : "password",
  "in" : "query",
  "description" : "user password.",
  "required" : true,
  "type" : "string",
  "format" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/DoneResponse"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/user/activate": {
    
    "post": {
      "summary": "",
      "description":"after signup we should send an activation code to user by sms and user send it back to us by this function",
      "x-swagger-router-controller": "User",
      "tags": ["User"],
      "operationId": "userActivatePost",
      "parameters": [
        {
  "name" : "phone_number",
  "in" : "query",
  "description" : "user phone number.",
  "required" : true,
  "type" : "number",
  "format" : "string"
},
        {
  "name" : "activation_code",
  "in" : "query",
  "description" : "activation code.",
  "required" : true,
  "type" : "number",
  "format" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/DoneResponse"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/user/signup": {
    
    "post": {
      "summary": "register user",
      "description":"register user by phone number and password (later we use phone number and verification code by sms like telegram)",
      "x-swagger-router-controller": "User",
      "tags": ["User"],
      "operationId": "userSignupPost",
      "parameters": [
        {
  "name" : "phone_number",
  "in" : "query",
  "description" : "user phone number.",
  "required" : true,
  "type" : "number",
  "format" : "string"
},
        {
  "name" : "password",
  "in" : "query",
  "description" : "user password.",
  "required" : true,
  "type" : "string",
  "format" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/DoneResponse"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/user/set_profile": {
    
    "post": {
      "summary": "save user Profile",
      "description":"همه داده‌های ارسالی ذخیره و جایگزین داده های قبلی می‌شوندو در صورتی که داده‌ای ارسال نشده یا خالی ارسال شده است در نظر گرفته نمی‌شود.",
      "x-swagger-router-controller": "User",
      "tags": ["User"],
      "operationId": "userSetProfilePost",
      "parameters": [
        {
  "name" : "full_name",
  "in" : "query",
  "description" : "full name of the user.",
  "required" : true,
  "type" : "string",
  "format" : "string"
},
        {
  "name" : "picture",
  "in" : "query",
  "description" : "Image avatar URL of the user.",
  "required" : true,
  "type" : "string",
  "format" : "string"
},
        {
  "name" : "username",
  "in" : "query",
  "description" : "username of user.",
  "required" : true,
  "type" : "string",
  "format" : "string"
},
        {
  "name" : "phone_number",
  "in" : "query",
  "description" : "user phone number.",
  "required" : true,
  "type" : "number",
  "format" : "string"
},
        {
  "name" : "password",
  "in" : "query",
  "description" : "user password.",
  "required" : true,
  "type" : "string",
  "format" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/DoneResponse"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
...
  "/place/get_places_list": {
    
    "get": {
      "summary": "get list of places",
      "description":"لیست مکان ها سمت کلاینت در دیتابیس ذخیره می‌شود پس در این تابع تاریخ آخرین بروزرسانی پایگاه داده ارسال می‌شود و همه مکانهای جدید پس از آن برای کلاینت ارسال می‌شود. اگه تاریخ ارسالی خالی بود همه مکانها جهت ذخیره در کلاینت ارسال خواهد شد.",
      "x-swagger-router-controller": "Place",
      "tags": ["Place"],
      "operationId": "placeGetPlacesListGet",
      "parameters": [
        {
  "name" : "last_update",
  "in" : "query",
  "description" : "last database update date.",
  "required" : true,
  "type" : "string",
  "format" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "An array of Places",
  "schema" : {
    "type" : "array",
    "items" : {
      "$ref" : "#/definitions/Place"
    }
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/place/get_places_details": {
    
    "get": {
      "summary": "get details of a place",
      "description":"به وسیله این تابع اطلاعات یک مکان به صورت کامل گرفته می‌شود، شامل ویژگی های خود مکان تعداد ووت و بررسی های مکان",
      "x-swagger-router-controller": "Place",
      "tags": ["Place"],
      "operationId": "placeGetPlacesDetailsGet",
      "parameters": [
        {
  "name" : "last_update",
  "in" : "query",
  "description" : "last database update date.",
  "required" : true,
  "type" : "string",
  "format" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/Place"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/place/add_new_place": {
    
    "post": {
      "summary": "add new place",
      "description":"اطلاعات یک مکان که فقط شامل نام و توضیحات و تصاویر آن مکان میباشد توسط کاربر ارسال شده و سمت سرور ذخیره می‌شود",
      "x-swagger-router-controller": "Place",
      "tags": ["Place"],
      "operationId": "placeAddNewPlacePost",
      "parameters": [
        {
  "name" : "name",
  "in" : "query",
  "required" : false,
  "type" : "string"
},
        {
  "name" : "description",
  "in" : "query",
  "required" : false,
  "type" : "string"
},
        {
  "name" : "images",
  "in" : "query",
  "required" : false,
  "type" : "string"
},
        {
  "name" : "user_id",
  "in" : "query",
  "required" : false,
  "type" : "number"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/DoneResponse"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/place/report_place": {
    
    "post": {
      "summary": "report a place",
      "description":"گزارش تخلف در یک مکان به کمک این تابع انجام می‌گیرد.",
      "x-swagger-router-controller": "Place",
      "tags": ["Place"],
      "operationId": "placeReportPlacePost",
      "parameters": [
        {
  "name" : "place_id",
  "in" : "query",
  "required" : false,
  "type" : "number"
},
        {
  "name" : "user_id",
  "in" : "query",
  "required" : false,
  "type" : "number"
},
        {
  "name" : "reason",
  "in" : "query",
  "description" : "دلیل گزارش دادن این مورد",
  "required" : false,
  "type" : "number"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/DoneResponse"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } ,
  "/place/rate_place": {
    
    "post": {
      "summary": "rate a place +1 or -1",
      "description":"امتیاز دهی به مکان توسط این تابع انجام می‌گیرد",
      "x-swagger-router-controller": "Place",
      "tags": ["Place"],
      "operationId": "placeRatePlacePost",
      "parameters": [
        {
  "name" : "place_id",
  "in" : "query",
  "required" : false,
  "type" : "number"
},
        {
  "name" : "user_id",
  "in" : "query",
  "required" : false,
  "type" : "number"
},
        {
  "name" : "vote",
  "in" : "query",
  "description" : "امتیاز مثبت یا منفی",
  "required" : false,
  "type" : "number"
}
        
      ],
      "responses": {
        "200": {
  "description" : "عملیات سرور به درستی انجام شده و یک پیام در قسمت message قرار می‌گیرد",
  "schema" : {
    "$ref" : "#/definitions/DoneResponse"
  }
}
        ,
        "default": {
  "description" : "Unexpected error",
  "schema" : {
    "$ref" : "#/definitions/Error"
  }
}
        
        
      }
    } 
    
  } 
  }, "definitions": {
  "User": {
  "type" : "object",
  "properties" : {
    "id" : {
      "type" : "string",
      "description" : "id."
    },
    "phone_number" : {
      "type" : "number",
      "description" : "phone number of the user."
    },
    "username" : {
      "type" : "string",
      "description" : "username of user"
    },
    "password" : {
      "type" : "string",
      "description" : "password of user"
    },
    "avatar" : {
      "type" : "string",
      "description" : "Image URL of the user."
    }
  }
},"Message": {
...
},"Place": {
  "type" : "object",
  "properties" : {
    "id" : {
      "type" : "number",
      "description" : "place id"
    },
    "name" : {
      "type" : "string",
      "description" : "name of the place."
    },
    "vote_number" : {
      "type" : "number",
      "description" : "number of vote for this place, can be positive or negative."
    },
    "images" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/PlaceImage"
      }
    },
    "reviews" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/Review"
      }
    }
  }
},"PlaceImage": {
  "type" : "object",
  "properties" : {
    "id" : {
      "type" : "number"
    },
    "place_id" : {
      "type" : "number"
    },
    "orginal_url" : {
      "type" : "string"
    },
    "thumbnail_url" : {
      "type" : "string"
    },
    "upload_date" : {
      "type" : "string"
    }
  }
},"Room": {
...
},"Review": {
  "type" : "object",
  "properties" : {
    "id" : {
      "type" : "number"
    },
    "user_id" : {
      "type" : "number"
    },
    "place_id" : {
      "type" : "number"
    },
    "text" : {
      "type" : "string"
    },
    "photo" : {
      "type" : "string"
    },
    "video" : {
      "type" : "string"
    },
    "status" : {
      "type" : "string",
      "description" : "status can be accepted, rejected, waiting"
    },
    "date" : {
      "type" : "string"
    }
  },
  "description" : "any user can send review for each place"
},"ReviewVote": {
...
},"Error": {
  "type" : "object",
  "properties" : {
    "code" : {
      "type" : "integer",
      "format" : "int32"
    },
    "message" : {
      "type" : "string"
    }
  }
},"DoneResponse": {
  "type" : "object",
  "properties" : {
    "code" : {
      "type" : "integer",
      "format" : "int32"
    },
    "message" : {
      "type" : "string"
    }
  }
}
  }
}